refactor: Legacy 코드 파일 전부 삭제
* Legacy 코드 파일 전부 삭제
