--- src/components/common/Title.tsx
@@ -1,10 +1,0 @@
-import { Txt } from './Txt';
-import { Wrapper } from './Wrapper';
-
-export default function Title({ title }: { title: string }) {
-  return (
-    <Wrapper className='title'>
-      <Txt typography='h4'>{title}</Txt>
-    </Wrapper>
-  );
-}
