--- src/constants/colors.ts
@@ -1,11 +1,0 @@
-const colors = {
-  white: '#ffffff',
-  gray01: '#2f3336',
-  gray02: '#71767b',
-  black: '#000000',
-  primary: '#1d9bf0',
-  red01: '#ff6347;',
-  green01: '#14b987',
-};
-
-export default colors;
